+++ js/src/mpl.js
@@ -14,19 +14,6 @@
 window.mpl = {};
 
 
-mpl.get_websocket_type = function() {
-    if (typeof(WebSocket) !== 'undefined') {
-        return WebSocket;
-    } else if (typeof(MozWebSocket) !== 'undefined') {
-        return MozWebSocket;
-    } else {
-        alert('Your browser does not have WebSocket support.' +
-              'Please try Chrome, Safari or Firefox ≥ 6. ' +
-              'Firefox 4 and 5 are also supported but you ' +
-              'have to enable WebSockets in about:config.');
-    };
-}
-
 mpl.figure = function(figure_id, websocket, ondownload, parent_element) {
     this.id = figure_id;
 
@@ -90,11 +77,20 @@ mpl.figure.prototype._init_header = function() {
 }
 
 mpl.figure.prototype._canvas_extra_style = function(canvas_div) {
-
+    // this is important to make the div 'focusable'
+    canvas_div.attr('tabindex', 0);
 }
 
 mpl.figure.prototype._root_extra_style = function(canvas_div) {
+    var fig = this;
+    canvas_div.on("remove", function(){
+        fig.close_ws(fig, {});
+    });
+}
 
+mpl.figure.prototype.close_ws = function(fig, msg){
+    fig.send_message('closing', msg);
+    fig.ws.close()
 }
 
 mpl.figure.prototype._init_canvas = function() {
@@ -223,9 +219,18 @@ mpl.figure.prototype._init_image = function() {
 mpl.figure.prototype._init_toolbar = function() {
     var fig = this;
 
-    var nav_element = $('<div/>')
-    nav_element.attr('style', 'width: 100%');
-    this.root.append(nav_element);
+    var toolbar_container = this.toolbar = $('<div class="jupyter-widgets widget-container widget-box widget-hbox"/>')
+    this.root.prepend(toolbar_container);
+
+    // Add the stop interaction button to the window.
+    var button = $('<button class="jupyter-widgets jupyter-button" href="#" title="Toggle Interaction"><i class="fa fa-bars"></i></button>');
+    button.attr('style', 'outline:none');
+    button.click(function (evt) { fig.toggle_interaction(fig, {}); } );
+    button.mouseover('Toggle Interaction', toolbar_mouse_event);
+    toolbar_container.append(button);
+
+    var toolbar = this.toolbar = $('<div class="jupyter-widgets widget-container widget-box widget-hbox"/>')
+    toolbar_container.append(toolbar);
 
     // Define a callback function for later on.
     function toolbar_event(event) {
@@ -240,58 +245,35 @@ mpl.figure.prototype._init_toolbar = function() {
         var tooltip = mpl.toolbar_items[toolbar_ind][1];
         var image = mpl.toolbar_items[toolbar_ind][2];
         var method_name = mpl.toolbar_items[toolbar_ind][3];
+        if (!name) { continue; };
 
-        if (!name) {
-            // put a spacer in here.
-            continue;
-        }
-        var button = $('<button/>');
-        button.addClass('ui-button ui-widget ui-state-default ui-corner-all ' +
-                        'ui-button-icon-only');
-        button.attr('role', 'button');
-        button.attr('aria-disabled', 'false');
+        var button = $('<button class="jupyter-widgets jupyter-button" href="#" title="' + name + '"><i class="fa ' + image + '"></i></button>');
+        button.attr('style', 'outline:none');
         button.click(method_name, toolbar_event);
         button.mouseover(tooltip, toolbar_mouse_event);
-
-        var icon_img = $('<span/>');
-        icon_img.addClass('ui-button-icon-primary ui-icon');
-        icon_img.addClass(image);
-        icon_img.addClass('ui-corner-all');
-
-        var tooltip_span = $('<span/>');
-        tooltip_span.addClass('ui-button-text');
-        tooltip_span.html(tooltip);
-
-        button.append(icon_img);
-        button.append(tooltip_span);
-
-        nav_element.append(button);
+        toolbar.append(button);
     }
 
-    var fmt_picker_span = $('<span/>');
-
-    var fmt_picker = $('<select/>');
-    fmt_picker.addClass('mpl-toolbar-option ui-widget ui-widget-content');
-    fmt_picker_span.append(fmt_picker);
-    nav_element.append(fmt_picker_span);
-    this.format_dropdown = fmt_picker[0];
-
-    for (var ind in mpl.extensions) {
-        var fmt = mpl.extensions[ind];
-        var option = $(
-            '<option/>', {selected: fmt === mpl.default_extension}).html(fmt);
-        fmt_picker.append(option)
-    }
-
-    // Add hover states to the ui-buttons
-    $( ".ui-button" ).hover(
-        function() { $(this).addClass("ui-state-hover");},
-        function() { $(this).removeClass("ui-state-hover");}
-    );
-
-    var status_bar = $('<span class="mpl-message"/>');
-    nav_element.append(status_bar);
+    // Add the status bar.
+    var status_bar = $('<span class="mpl-message" style="text-align:right; float: right;"/>');
+    toolbar.append(status_bar);
     this.message = status_bar[0];
+}
+
+mpl.figure.prototype.toggle_interaction = function(fig, msg) {
+    // Toggle the interactivity of the figure.
+    var visible = fig.toolbar.is(':visible');
+    if (visible) {
+        fig.toolbar.hide();
+        fig.canvas_div.hide();
+        fig.imageObj_div.show();
+        fig.imageObj.style.width = fig.canvas.style.width;
+        fig.imageObj.style.height = fig.canvas.style.height;
+    } else {
+        fig.toolbar.show();
+        fig.canvas_div.show();
+        fig.imageObj_div.hide();
+    }
 }
 
 mpl.figure.prototype.request_resize = function(x_pixels, y_pixels) {
@@ -315,9 +297,7 @@ mpl.figure.prototype.send_draw_message = function() {
 
 
 mpl.figure.prototype.handle_save = function(fig, msg) {
-    var format_dropdown = fig.format_dropdown;
-    var format = format_dropdown.options[format_dropdown.selectedIndex].value;
-    fig.ondownload(fig, format);
+    fig.ondownload(fig, null);
 }
 
 
@@ -388,7 +368,10 @@ mpl.figure.prototype.handle_image_mode = function(fig, msg) {
 }
 
 mpl.figure.prototype.updated_canvas_event = function() {
-    // Called whenever the canvas gets updated.
+    // Tell Jupyter that the notebook contents must change.
+    if (window.Jupyter) {
+        Jupyter.notebook.set_dirty(true);
+    }
     this.send_message("ack", {});
 }
 
@@ -504,10 +487,6 @@ mpl.figure.prototype.mouse_event = function(event, name) {
     return false;
 }
 
-mpl.figure.prototype._key_event_extra = function(event, name) {
-    // Handle any extra behaviour associated with a key event
-}
-
 mpl.figure.prototype.key_event = function(event, name) {
 
     // Prevent repeat events
@@ -531,8 +510,6 @@ mpl.figure.prototype.key_event = function(event, name) {
 
     value += 'k';
     value += event.which.toString();
-
-    this._key_event_extra(event, name);
 
     this.send_message(name, {key: value,
                              guiEvent: simpleKeys(event)});
